+++ src/errors/transport/mod.rs
@@ -8,6 +8,7 @@ use thiserror::Error;
 use crate::errors::transport::{tcp::TcpError, udp::UdpError};
 
 pub mod icmp;
+pub mod icmpv6;
 pub mod tcp;
 pub mod udp;