+++ src/parse/transport/protocols/mod.rs
@@ -8,6 +8,7 @@ use serde::Serialize;
 use crate::Transport;
 
 pub mod icmp;
+pub mod icmpv6;
 pub mod tcp;
 pub mod udp;